implemented: AVL_tree memory management. the balancing functions of the AVL_tree. (rolls) AVL_tree.add().
TODO:
download new .cpp file from the course staff and replace what data we have there.
comparison operators.
decide on pointer architecture for the nodes (e.g shared pointers, 1 unique_ptr in the big tree and regular pointers for the others, etc).
the rest of the AVL_tree public functions.

notes:
the player and team files are not updated. they run but the code there needs to be checked before its used.